requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Deprecated==1.2.15
Django==5.1.3
django-bootstrap5==24.3
djangorestframework==3.15.2
drf-standardized-errors==0.14.1
googleapis-common-protos==1.66.0
grpcio==1.68.0
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
kombu==5.4.2
numpy==2.2.0
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-asgi==0.49b2
opentelemetry-instrumentation-asyncio==0.49b2
opentelemetry-instrumentation-celery==0.49b2
opentelemetry-instrumentation-dbapi==0.49b2
opentelemetry-instrumentation-django==0.49b2
opentelemetry-instrumentation-grpc==0.49b2
opentelemetry-instrumentation-logging==0.49b2
opentelemetry-instrumentation-pika==0.49b2
opentelemetry-instrumentation-redis==0.49b2
opentelemetry-instrumentation-requests==0.49b2
opentelemetry-instrumentation-sqlite3==0.49b2
opentelemetry-instrumentation-threading==0.49b2
opentelemetry-instrumentation-urllib==0.49b2
opentelemetry-instrumentation-urllib3==0.49b2
opentelemetry-instrumentation-wsgi==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pika==1.3.2
pluggy==1.5.0
prompt_toolkit==3.0.48
protobuf==5.28.3
psycopg==3.2.3
psycopg-binary==3.2.3
PySocks==1.7.1
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.0
requests==2.32.3
selenium==4.26.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
wrapt==1.17.0
wsproto==1.2.0
zipp==3.21.0
